refactor out to useKeybinds
